always check remote server address
The remote server address is only checked when the user press the
validation button, and the remote server configuration can be
saved without address, this can cause a NullPointerException.
The validation should be always done, and the user should know that if
the adress is not configured, it must be overriden in the job
configuration.